Fix bugs
Fix bugs in maximum size for some integers
Fix relashionship issues between 'piloto', 'voo', 'aluno',
'instrutor' and 'voo_supervisionado'

+++ app/main/model/aluno.py
@@ -11,8 +11,8 @@ class Aluno(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     nome = db.Column(db.String(80), unique=False, nullable=False)
-    CPF = db.Column(db.Integer, unique=True, nullable=False)
-    matricula = db.Column(db.Integer, unique=True, nullable=False)
+    CPF = db.Column(db.BigInteger, unique=True, nullable=False)
+    matricula = db.Column(db.BigInteger, unique=True, nullable=False)
     email = db.Column(db.String(120), unique=True, nullable=False)
     senha = db.Column(db.String(32), unique=False, nullable=False)
     horas_de_voo = db.Column(db.Float, unique=False, nullable=False)
